+++ src/js_modules/profile/Suscriptions.jsx
@@ -1,0 +1,11 @@
+const Suscriptions = () => {
+  console.log('Suscriptions');
+
+  return (
+    <div className="suscriptions">
+      <h1>Suscriptions</h1>
+    </div>
+  );
+};
+
+export default Suscriptions;
